# openapi/admin-api.yaml
openapi: 3.1.0
info:
  title: Leonettis Admin API
  version: 1.0.0
  description: |
    Admin-only APIs for authentication, listings CRUD, media storage, and one-time seeding.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/admin/listings:
    get:
      tags: [Admin Listings]
      summary: Get all listings (admin view)
      security:
        - adminCookie: []
      responses:
        '200':
          description: Listing collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listing'
        '401':
          description: Unauthorized
    post:
      tags: [Admin Listings]
      summary: Create listing
      security:
        - adminCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingInput'
      responses:
        '201':
          description: Created listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  listing:
                    $ref: '#/components/schemas/Listing'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
...
  /api/admin/translate:
    post:
      tags: [Admin Translation]
      summary: Translate listing text across locales
      description: Translates title or description from one source locale into all remaining admin locales.
      security:
        - adminCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTranslateRequest'
      responses:
        '200':
          description: Translation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTranslateResponse'
        '400':
          description: Invalid request payload
        '401':
          description: Unauthorized
        '502':
          description: Invalid model output
        '500':
          description: Translation provider failure
...
components:
  securitySchemes:
    adminCookie:
      type: apiKey
      in: cookie
      name: admin_token
  schemas:
    ListingImage:
      type: object
      required: [url, name]
      properties:
        url:
          type: string
        name:
          type: string
        key:
          type: string
        description:
          type: string
...
    ListingVideo:
      type: object
      required: [url, name]
      properties:
        url:
          type: string
        name:
          type: string
        key:
          type: string
        description:
          type: string
        contentType:
          type: string
          description: Optional MIME type for uploaded videos (for example video/webm).
...
    AdminTranslateRequest:
      type: object
      required: [text, sourceLocale, field]
      properties:
        text:
          type: string
        sourceLocale:
          $ref: '#/components/schemas/Locale'
        field:
          type: string
          enum: [title, description]
    AdminTranslateResponse:
      type: object
      required: [translations]
      properties:
        translations:
          type: object
          description: Contains all locales except the provided sourceLocale.
          additionalProperties:
            type: string
...
    Locale:
      type: string
      enum: [en, fr, gr, de, it]
    ListingInput:
      type: object
      required:
        [title, slug, address, listingType, category, propertyType, tags]
      properties:
        id:
          type: string
        title:
          type: object
          additionalProperties:
            type: string
        description:
          type: object
          additionalProperties:
            type: string
        slug:
          type: string
        address:
          type: object
          properties:
            streetNumber:
              type: string
            streetName:
              type: string
            city:
              type: string
            state:
              type: string
            region:
              type: string
            zipCode:
              type: string
            country:
              type: string
            displayAddress:
              type: string
            coordinates:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
        listingType:
          type: string
          enum: [buy, rent]
        category:
          type: array
          items:
            type: string
        propertyType:
          type: string
        price:
          type: number
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareMetersInterior:
          type: number
        squareMetersOutdoor:
          type: number
        squareMetersTotal:
          type: number
        images:
          type: array
          items:
            $ref: '#/components/schemas/ListingImage'
        mainImage:
          type: string
        videos:
          type: array
          items:
            $ref: '#/components/schemas/ListingVideo'
        features:
          type: array
          description: Predefined feature IDs plus optional custom admin-defined values.
          items:
            type: string
        furnishing:
          type: string
        amenities:
          type: array
          description: Predefined amenity IDs plus optional custom admin-defined values.
          items:
            type: string
        suitableFor:
          type: array
          description: Predefined suitability IDs plus optional custom admin-defined values.
          items:
            type: string
        view:
          type: array
          description: Predefined view IDs plus optional custom admin-defined values.
          items:
            type: string
        publishedAt:
          type: string
        updatedAt:
          type: string
        status:
          type: string
          enum: [active, sold, rented, pending, under_offer]
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        urgent:
          type: boolean
        condition:
          type: string
        yearBuilt:
          type: number
        energyRating:
          type: string
        yearRenovated:
          type: number
        availableNow:
          type: boolean
        availableUponRequest:
          type: boolean
        availableFrom:
          type: string
        availableTo:
          type: string
        leaseDuration:
          type: number
        leaseDurationUnit:
          type: string
          enum: [month, year]
        leaseDurationType:
          type: string
          enum: [fixed, flexible]
    Listing:
      allOf:
        - $ref: '#/components/schemas/ListingInput'
        - type: object
          required: [id, slug, publishedAt, updatedAt]